% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMPLN.R
\name{compute_invSigma_mix_glasso}
\alias{compute_invSigma_mix_glasso}
\title{A function to}
\usage{
compute_invSigma_mix_glasso(wij, lambda, K, invSigma_init, penalty)
}
\value{
invSigma
}
\description{
A function to
}
