% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMPLN.R
\name{compute_invSigma_mix_huge}
\alias{compute_invSigma_mix_huge}
\title{A function to}
\usage{
compute_invSigma_mix_huge(wij, data, K, invSigma_init, penalty)
}
\arguments{
\item{data}{}
}
\value{
invSigma
}
\description{
A function to
}
