% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMPLN.R
\name{compute_lambdaij_mix}
\alias{compute_lambdaij_mix}
\title{A function to}
\usage{
compute_lambdaij_mix(X, invSigma, mu, lambda, k)
}
\arguments{
\item{invSigma}{}

\item{lambda}{}
}
\value{
lambdanew
}
\description{
A function to
}
