% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMPLN.R
\name{compute_ldensity}
\alias{compute_ldensity}
\title{A function to}
\usage{
compute_ldensity(X, invSigma, lambda, ldt, mu)
}
\arguments{
\item{X}{}

\item{invSigma}{}
}
\value{
val
}
\description{
A function to
}
