% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{A function to generate the synthetic data}
\usage{
generate(K, N, d, sp, type)
}
\arguments{
\item{K}{number of componenet in the synthetic data}

\item{N}{number of samples in the synthetic sample-taxa count matrix}

\item{d}{number of taxa in the synthetic sample-taxa count matrix}

\item{sp}{sparsity level. value between 0 and 1.}

\item{type}{"orig" original count data, "samp" sampled data , "TMM" after TMM normalization}
}
\description{
A function to generate the synthetic data
}
