% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMPLN.R
\name{MixMPLN}
\alias{MixMPLN}
\title{MixMPLN main}
\usage{
MixMPLN(X, K, penalty, init, rep)
}
\arguments{
\item{K}{number of components}

\item{penalty}{"no_sparsity", "CV", "StARS" , "fixed" , "iterative"}

\item{init}{"Kmeans"}

\item{rep}{number of repeat}

\item{out}{"precision" , "partial" , "adj"}

\item{threshold}{a value between 0 and 1 which indicates the threshod to generate the adjacency matrix}
}
\description{
This function receives the sample-taxa count matrix and cluser the samples based on the MixMPLN method.
}
